Refactor: Decompose TvShowDetailsScreen into smaller components
This commit refactors the `TvShowDetailsScreen` by extracting its UI logic into smaller, reusable, and more manageable composable components. The monolithic screen composable has been broken down into multiple focused components, each responsible for a specific section of the UI.

This improves code organization, readability, and maintainability within the `tv-shows` feature module.

Key changes:
- **Decomposition of `TvShowDetailsScreen.kt`:** The main screen composable now primarily orchestrates the layout of the various section components. All UI logic for individual sections has been moved out.
- **Creation of a `components` package:** A new package `features/tv-shows/src/commonMain/kotlin/.../ui/tv_show_details/components` has been created to house the new composables.
- **Extracted Components:** The following composables have been extracted into their own files:
    - `HeroSection`: Displays the backdrop, poster, title, and rating.
    - `BasicInfoSection`: Shows core details like air dates, season/episode counts, and language.
    - `OverviewSection`: Displays the show's overview and tagline.
    - `RatingsSection`: Shows cards for vote average, vote count, and popularity.
    - `SeriesInfoSection`: Displays status, production state, and homepage link.
    - `ProductionSection`: Lists creators and production companies/countries.
    - `EpisodesSection`: Details for the last and next episodes to air.
    - `GenresSection`: Displays genre chips.
    - `NetworksSection`: Displays network chips.
    - `LanguagesSection`: Lists spoken and available languages.
    - `CastSection` & `CastMemberCard`: For displaying the cast list.
    - `TrailersSection` & `TrailerCard`: For displaying video trailers.
    - `InfoRow` & `RatingCard`: Reusable utility composables for displaying data points.
